refactor(ontario-summary-list): extract types and consts to companion types file
